+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.5.0
 
+* NewApplication function and  Start function of the Application interface were replaced by StartApp function
 * Added HasFocus function to the View interface
 * Added the UserAgent function to the Session interface
 * Added the following properties to TableView: "selection-mode", "allow-selection", "current", "current-style", "current-inactive-style"